+++ example/flutter/github_search/lib/main.dart
@@ -1,33 +1,22 @@
 import 'package:flutter/material.dart';
-import 'package:github_search/github_search_api.dart';
-import 'package:github_search/github_search_widget.dart';
+import 'package:github_search/github_api.dart';
+import 'package:github_search/search_widget.dart';
 
-void main() {
-  runApp(new RxDartGithubSearchApp(
-    api: new GithubApi(),
-  ));
+void main(GithubApi api) {
+  runApp(new SearchApp(api: api));
 }
 
-class RxDartGithubSearchApp extends StatefulWidget {
+class SearchApp extends StatefulWidget {
   final GithubApi api;
 
-  RxDartGithubSearchApp({Key key, this.api}) : super(key: key);
+  SearchApp({Key key, this.api}) : super(key: key);
 
   @override
   _RxDartGithubSearchAppState createState() =>
       new _RxDartGithubSearchAppState();
 }
 
-class _RxDartGithubSearchAppState extends State<RxDartGithubSearchApp> {
-  SearchBloc bloc;
-
-  @override
-  void initState() {
-    super.initState();
-
-    bloc = new SearchBloc(widget.api);
-  }
-
+class _RxDartGithubSearchAppState extends State<SearchApp> {
   @override
   Widget build(BuildContext context) {
     return new MaterialApp(
@@ -36,10 +25,7 @@ class _RxDartGithubSearchAppState extends State<RxDartGithubSearchApp> {
         brightness: Brightness.dark,
         primarySwatch: Colors.grey,
       ),
-      home: new SearchScreen(
-        onTextChanged: bloc.onTextChanged.add,
-        state: bloc.state,
-      ),
+      home: new SearchScreen(api: widget.api),
     );
   }
 }
