+++ example/flutter/github_search/pubspec.yaml
@@ -6,7 +6,11 @@ dependencies:
     sdk: flutter
   rxdart:
     path: ../../../
-  flutter_stream_friends: 0.5.0
+
+dev_dependencies:
+  mockito: 2.2.3
+  flutter_test:
+    sdk: flutter
 
 # For information on the generic Dart part of this file, see the
 # following page: https://www.dartlang.org/tools/pub/pubspec
